rename network architectures
